requirements.txt
Pillow==11.1.0
mss==10.0.0
pynput==1.8.1
pystray==0.19.5